Refactor: apis 디렉토리 생성
api 관련 로직들을 apis 디렉토리에 모았습니다.
